change file name project2 to app
